chore: styled => highlight and separate into another file
